requirements.txt
asgiref==3.8.1
coverage==7.6.10
Django==5.1.5
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
exceptiongroup==1.2.2
factory-boy==3.3.0
Faker==19.3.0
iniconfig==2.0.0
packaging==24.2
phonenumbers==8.13.53
pluggy==1.5.0
PyJWT==2.10.1
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-mock==3.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
